+++ hpOneView/resources/networking/ethernet_networks.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -31,10 +31,11 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import (Resource, ResourcePatchMixin,
+                                          ensure_resource_client)
 
 
-class EthernetNetworks(object):
+class EthernetNetworks(ResourcePatchMixin, Resource):
     """
     Ethernet Networks API client.
 
@@ -45,92 +46,19 @@ class EthernetNetworks(object):
         '200': {"type": "ethernet-networkV3"},
         '300': {"type": "ethernet-networkV300"},
         '500': {"type": "ethernet-networkV300"},
-        '600': {"type": "ethernet-networkV4"}
+        '600': {"type": "ethernet-networkV4"},
+        '800': {"type": "ethernet-networkV4"}
     }
     BULK_DEFAULT_VALUES = {
         '200': {"type": "bulk-ethernet-network"},
         '300': {"type": "bulk-ethernet-network"},
         '500': {"type": "bulk-ethernet-network"},
-        '600': {"type": "bulk-ethernet-networkV1"}
+        '600': {"type": "bulk-ethernet-networkV1"},
+        '800': {"type": "bulk-ethernet-networkV1"},
     }
 
-    def __init__(self, con):
-        self._connection = con
-        self._client = ResourceClient(con, self.URI)
-
-    def get_all(self, start=0, count=-1, filter='', sort=''):
-        """
-        Gets a paginated collection of Ethernet networks. The collection is based on optional sorting and filtering
-        and is constrained by start and count parameters.
-
-        Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
-
-        Returns:
-            list: A list of ethernet networks.
-        """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
-
-    def delete(self, resource, force=False, timeout=-1):
-        """
-        Deletes an Ethernet network.
-
-        Any deployed connections that are using the network are placed in the 'Failed' state.
-
-        Args:
-            resource: dict object to delete
-            force:
-                 If set to true, the operation completes despite any problems with
-                 network connectivity or errors on the resource itself. The default is false.
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView; it just stops waiting for its completion.
-
-        Returns:
-            bool: Indicates if the resource was successfully deleted.
-
-        """
-        return self._client.delete(resource, force=force, timeout=timeout)
-
-    def get(self, id_or_uri):
-        """
-        Gets the Ethernet network.
-
-        Args:
-            id_or_uri: ID or URI of Ethernet network.
-
-        Returns:
-            dict: The ethernet network.
-        """
-        return self._client.get(id_or_uri)
-
-    def create(self, resource, timeout=-1):
-        """
-        Creates an Ethernet network.
-
-        Args:
-            resource (dict): Object to create.
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView; it just stops waiting for its completion.
-
-        Returns:
-            dict: Created resource.
-
-        """
-        return self._client.create(resource, timeout=timeout, default_values=self.DEFAULT_VALUES)
+    def __init__(self, connection, data=None):
+        super(EthernetNetworks, self).__init__(connection, data)
 
     def create_bulk(self, resource, timeout=-1):
         """
@@ -147,7 +75,10 @@ class EthernetNetworks(object):
 
         """
         uri = self.URI + '/bulk'
-        self._client.create(resource, uri=uri, timeout=timeout, default_values=self.BULK_DEFAULT_VALUES)
+        default_values = self._get_default_values(self.BULK_DEFAULT_VALUES)
+        updated_data = self._helper.update_resource_fields(resource, default_values)
+
+        self._helper.create(updated_data, uri=uri, timeout=timeout)
 
         return self.get_range(resource['namePrefix'], resource['vlanIdRange'])
 
@@ -219,36 +150,8 @@ class EthernetNetworks(object):
 
         return vlan_ids
 
-    def update(self, resource, timeout=-1):
-        """
-        Updates an Ethernet network.
-
-        Args:
-            resource: dict object to update
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView; it just stops waiting for its completion.
-
-        Returns: Updated resource.
-
-        """
-        return self._client.update(resource, timeout=timeout, default_values=self.DEFAULT_VALUES)
-
-    def get_by(self, field, value):
-        """
-        Gets all Ethernet networks that match the filter.
-        The search is case-insensitive.
-
-        Args:
-            field: field name to filter
-            value: value to filter
-
-        Returns:
-            list: A list of ethernet networks.
-        """
-        return self._client.get_by(field, value)
-
-    def get_associated_profiles(self, id_or_uri):
+    @ensure_resource_client
+    def get_associated_profiles(self):
         """
         Gets the URIs of profiles which are using an Ethernet network.
 
@@ -259,38 +162,17 @@ class EthernetNetworks(object):
             list: URIs of the associated profiles.
 
         """
-        uri = self._client.build_uri(id_or_uri) + "/associatedProfiles"
-        return self._client.get(uri)
+        uri = "{}/associatedProfiles".format(self.data['uri'])
+        return self._helper.do_get(uri)
 
-    def get_associated_uplink_groups(self, id_or_uri):
+    @ensure_resource_client
+    def get_associated_uplink_groups(self):
         """
         Gets the uplink sets which are using an Ethernet network.
-
-        Args:
-            id_or_uri: Can be either the logical interconnect group id or the logical interconnect group uri
 
         Returns:
             list: URIs of the associated uplink sets.
 
         """
-        uri = self._client.build_uri(id_or_uri) + "/associatedUplinkGroups"
-        return self._client.get(uri)
-
-    def patch(self, id_or_uri, operation, path, value, timeout=-1):
-        """
-        Uses the PATCH to update the given resource.
-
-        Only one operation can be performed in each PATCH call.
-
-        Args:
-            id_or_uri: Can be either the resource ID or the resource URI.
-            operation: Patch operation
-            path: Path
-            value: Value
-            timeout: Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView; it just stops waiting for its completion.
-
-        Returns:
-            dict: Updated resource.
-        """
-        return self._client.patch(id_or_uri, operation, path, value, timeout=timeout)
+        uri = "{}/associatedUplinkGroups".format(self.data['uri'])
+        return self._helper.do_get(uri)