+++ CHANGELOG.md
@@ -27,10 +27,18 @@ Resource data will be available with the resource object. This enhancement helps
 - Connection template
 - Enclosure
 - Enclosure group
+- Ethernet network
 - FC network
 - FCOE network
 - Interconnect type
+- Internal link set
+- Managed SAN
+- SAS interconnect
+- SAS interconnect type
 - SAS logical interconnect
+- Server hardware type
+- Switch type
+- Uplink set
 
 # 4.8.0
 #### Notes